Add ContentType enum and ImageResponse class***
The developer added a new file `rossa/types.py` which includes the `ContentType` enum class. This enum class defines different content types such as image, video, audio, text, and three-dimensional.

In addition, the developer modified the `workflows/image_object_removal_lama.py` file to import the `ContentType` enum and the `ImageResponse` class. The `download` method was also modified to assign an instance of `SimpleLama` to `self.simple_lama`. The `load` method was modified to assign an instance of `SimpleLama` to `self.simple_lama`. The `Workflow` class was modified to return an `ImageResponse` object in the `run` method.

The `rossa/__init__.py` file was modified to import the `ContentType` enum and the `ImageResponse` class.

The `rossa/workflow.py` file was modified to import the `ContentType` enum and the `Option` class from the appropriate modules.